laying more groundwork for multi place

+++ LearningMono/Config/PlayerConfig.fs
@@ -9,7 +9,7 @@ open GameConfig
 type PlayerConfig =
     { BigMaxVelocity: float32
       SmallMaxVelocity: float32
-      MultiPlaceDelayMs: float32
+      MultiPlaceDelayMs: int64
       Acc: float32
       Slow: float32
       AABBConfig: AABB }
@@ -17,7 +17,7 @@ type PlayerConfig =
 let playerConfig =
     { BigMaxVelocity = 2f
       SmallMaxVelocity = 4.5f
-      MultiPlaceDelayMs = 400f
+      MultiPlaceDelayMs = 400
       Acc = 30f
       Slow = 40f
       AABBConfig =
